project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "itsAyush04",email = "135616593+itsAyush04@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pytest = "^8.4.1"

